requirements.txt
numpy==1.26.4
pandas==2.2.2
langgraph>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
python-dotenv==1.0.1
langchain-community>=0.2.0,<0.3.0
langchain-ollama==0.1.3
tqdm==4.67.3
pypdf==4.2.0
urllib3==1.26.18
neo4j==6.1.0
networkx==3.6.1
pyvis==0.3.2